1) Added support for MD5 hashed passwords in ProtoMUCK database by using a     new database flag, DB_MD5PASSES (0x04), which is set automatically when     converting from an FB6 database or when you use -pwconvert. 2) Added -pwconvert command-line option to convert current passwords to MD5     format.  Since databases default to the old method of clear-text     passwords, if you want them to be MD5 hashes, you must run this at     least once, but it is safe to run multiple times on the same database. 3) Increased version to 2.0b7.02.

+++ trunk/proto2/src/random.c
@@ -25,7 +25,7 @@
  * will fill a supplied 16-byte array with the digest.
  */
 
-typedef unsigned long word32;
+typedef unsigned int word32;    /* this used to be long, but that breaks on 64-bit machines. -hinoserm */
 typedef unsigned char byte;
 
 struct xMD5Context {
@@ -38,7 +38,6 @@ void xMD5Init(struct xMD5Context *context);
 void xMD5Update(struct xMD5Context *context, byte const *buf, int len);
 void xMD5Final(byte digest[16], struct xMD5Context *context);
 void xMD5Transform(word32 buf[4], word32 const in[16]);
-
 
 /*
  * Shuffle the bytes into little-endian order within words, as per the
